add auth decorator, login_required

+++ base/views.py
@@ -1,5 +1,7 @@
+from django.http import HttpResponse
 from django.shortcuts import render, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
@@ -65,11 +67,15 @@ def home(request):
     context = {'rooms': rooms, 'topics': topics, 'rooms_count': rooms_count}
     return render(request, 'base/home.html', context)
 
+
+
 def room(request, pk):
     room = Room.objects.get(id=pk)
     context = {'room': room}
     return render(request, 'base/room.html', context)
 
+
+@login_required(login_url='/login')
 def createRoom(request):
     """create room using room form with fields: host, topic, name, description"""
     form = RoomForm()
@@ -83,10 +89,15 @@ def createRoom(request):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def updateRoom(request, pk):
     """Update room details"""
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room) # instance=room to prefill
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
 
     if request.method == "POST":
         form = RoomForm(request.POST, instance=room) # instance=room to tell which room
@@ -97,9 +108,15 @@ def updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+
+@login_required(login_url='/login')
 def deleteRoom(request, pk):
     """Delete a room"""
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host:
+        return HttpResponse('You are not allowed here!!')
+
     if request.method == 'POST':
         room.delete()
         return redirect('home')
